gcc_mt_m12_v2/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
pytest==7.4.4

# Logging and utilities
structlog==24.1.0

# AWS SDK - SERVICE: AWS_S3 (detected from script Section 4)
boto3==1.34.0
botocore==1.34.0

# Caching (Redis)
redis==5.0.1

# Database (PostgreSQL for audit logging)
psycopg2-binary==2.9.9

# Jupyter notebook support
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.28.0

# Data handling
pandas==2.1.4

# Type checking (development)
mypy==1.8.0
types-redis==4.6.0.11
